+++ core/src/main/scala/kafka/log/LogSegment.scala
@@ -59,7 +59,7 @@ class LogSegment(val messageSet: FileMessageSet,
       // append the messages
       messageSet.append(messages)
       updateFirstAppendTime()
-      this.bytesSinceLastIndexEntry += messages.sizeInBytes.toInt
+      this.bytesSinceLastIndexEntry += messages.sizeInBytes
     }
   }
